refactor: reorganize category group types and utility functions
